day-planner/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
alembic
pydantic>=2.0
httpx
python-dotenv
itsdangerous
